requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
dnspython==2.8.0
Flask==3.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
